refactor: Modified color wheel when selecting colors
The selection is done as usual, but the selected color is marked
with two strokes, and inner and an outter one. The gap between
the wheel and the border is just the dimension of the outter
stroke.

The name of the renderer has been changed to identify it as an
arc renderer. Also a circle renderer it has been defined but
not implemented yet.+